chore: prevent geometry data source response being cloned to avoid resource drain on large orgs
